Move OnThisDayExploreCollectionViewCell to own file.
